+++ src/plugins/intel_gpu/src/plugin/ops/interpolate.cpp
@@ -69,7 +69,7 @@ static cldnn::resample_type GetResampleType(ngraph::op::v4::Interpolate::Interpo
 static cldnn::resample::resample_axis GetInterpolationAxis(int32_t axis, uint32_t sz) {
     if (axis < 0)
         axis += sz;
-    if (axis < 0 || axis >= sz)
+    if (axis < 0 || axis >=  static_cast<int32_t>(sz))
         IE_THROW() << "Interpolate axis is not correspond to number of dimensions";
 
     // Difference in dimension ordering between IE and GPU plugin,
@@ -148,7 +148,7 @@ static void CreateInterpolateOp(Program& p, const std::shared_ptr<ngraph::op::v4
             axes.push_back(GetInterpolationAxis(axis, inputRank));
         }
     } else {
-        for (int i = 0; i < inputRank; ++i) {
+        for (size_t i = 0; i < inputRank; ++i) {
             axes.push_back(GetInterpolationAxis(i, inputRank));
         }
     }
