+++ src/plugins/intel_gpu/src/plugin/ops/scatter_nd_update.cpp
@@ -25,8 +25,8 @@ static void CreateScatterNDUpdateOp(Program& p, const std::shared_ptr<ngraph::op
         auto indices_last_dim = op->get_input_shape(1)[indices_rank - 1];
         auto data_shape = op->get_input_shape(0);
         bool valid = true;
-        for (int i = 0; i < indices.size(); ++i) {
-            if (indices[i] >= data_shape[i % indices_last_dim])
+        for (size_t i = 0; i < indices.size(); ++i) {
+            if (indices[i] >= static_cast<int>(data_shape[i % indices_last_dim]))
                 valid = false;
         }
